+++ src/platforms/wordpress/classes/Gantry/Framework/Document.php
@@ -29,11 +29,15 @@ class Document extends HtmlDocument
     /** @var array */
     protected static $availableFrameworks = [
         'jquery' => 'registerJquery',
+        'jquery.2' => 'registerJquery',
+        'jquery.3' => 'registerJquery3',
         'jquery.framework' => 'registerJquery',
         'jquery.ui.core' => 'registerJqueryUiCore',
         'jquery.ui.sortable' => 'registerJqueryUiSortable',
         'bootstrap.2' => 'registerBootstrap2',
         'bootstrap.3' => 'registerBootstrap3',
+        'bootstrap.4' => 'registerBootstrap4',
+        'bootstrap.5' => 'registerBootstrap5',
         'mootools' => 'registerMootools',
         'mootools.framework' => 'registerMootools',
         'mootools.core' => 'registerMootools',
@@ -198,6 +202,11 @@ class Document extends HtmlDocument
         \wp_enqueue_script('jquery');
     }
 
+    protected static function registerJquery3()
+    {
+        \wp_enqueue_script('jquery', 'https://code.jquery.com/jquery-3.6.0.min.js');
+    }
+
     protected static function registerJqueryUiCore()
     {
         \wp_enqueue_script('jquery-ui-core');
@@ -210,12 +219,28 @@ class Document extends HtmlDocument
 
     protected static function registerBootstrap2()
     {
+        static::registerJquery();
+
         \wp_enqueue_script('bootstrap', 'https://maxcdn.bootstrapcdn.com/twitter-bootstrap/2.3.2/js/bootstrap.min.js');
     }
 
     protected static function registerBootstrap3()
     {
-        \wp_enqueue_script('bootstrap', 'https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js');
+        static::registerJquery();
+
+        \wp_enqueue_script('bootstrap', 'https://cdn.jsdelivr.net/npm/bootstrap@3.4.1/dist/js/bootstrap.min.js');
+    }
+
+    protected static function registerBootstrap4()
+    {
+        static::registerJquery3();
+
+        \wp_enqueue_script('bootstrap', 'https://cdn.jsdelivr.net/npm/bootstrap@4.6.0/dist/js/bootstrap.bundle.min.js');
+    }
+
+    protected static function registerBootstrap5()
+    {
+        \wp_enqueue_script('bootstrap', 'https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js');
     }
 
     protected static function registerMootools()