+++ CHANGELOG.md
@@ -29,6 +29,7 @@
         - Fixed infinite recursive loop if the menu contains menu particle (#2527)
         - Fixed `The menu selected is empty!` instead of displaying error message (#2806)
         - Fixed menu dropdown title
+        - Hydrogen/Helium: Fixed broken SCSS syntax
 1. [Joomla](#joomla)
     1. [](#new)
         - Updated minimum requirement to Joomla! 3.9 (includes themes)